requirements.txt
moviepy
numpy
Flask
Pillow
python-dotenv
openai
requests